requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
apturl==0.5.2
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
executing==2.2.0
frozenlist==1.4.1
git-filter-repo==2.47.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jeepney==0.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
matplotlib-inline==0.1.7
more-itertools==8.10.0
multidict==6.0.5
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==2.2.5
oauthlib==3.2.0
olefile==0.46
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==3.12.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycups==2.0.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.4.0
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
stack-data==0.6.3
sumolib==1.23.1
systemd-python==234
tornado==6.4.2
tqdm==4.67.1
traci==1.23.1
traitlets==5.14.3
twilio==9.2.4
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
yarl==1.9.11
zipp==1.0.0
